Added windows and linux deploy script
Linux is distributed to as a tar.gz where the user has to install SDL2.
Windows is distributed to as a zip including SDL2.dll. Changed directory
name gen to textures in order to make a prettier distribution as well as
be able to run twinpin with 'stack run' without too much hassle.
app/Paths_twinpin.hs will ensure that the textures in textures/* are
used