Changed to fullscreen mode
The game can be drawn in any aspect ratio or resolution. When the screen
has a different aspect ratio than 800/600, a frame is drawn. Optimized
the drawing of static images, such as shots, pillars or the frame since
rendering the frame from vector graphics for each frame was slow. Now
these images are only rendered once at the beginning. Extracted
responsibilities from Game to two new modules Menu and Visual